allow user to exit the channel if we show a PIN prompt on startup
If the current user is protected, we show a user list and PIN entry
dialog during startup. Clicking back or the close button caused an
infinite loop of this process, not allowing the user to exit without
pressing the remote HOME button. We now show "Exit" instead of "Close"
on the user selection, and will properly exit the channel on selection
or back button.