fix unclosable dialog and multi dialog prompts
We sometimes pop on activate when we call ShowPlaybackError, so we'll
need need to show a blocking dialog (which doesn't block the message
port anymore). The video error message for indirect media has been
removed when constructing a video item, and placed in
ShowPlaybackError. Now we'll only prompt the user once about the error,
instead showing a error dialog 3 times.