fix hiding or showing placeholder for empty rows
This is a more simplistic and less intrusive fix. The issue stems from a
server having multiple addresses (public/local+), in which more than one
are unreachable. The pending request, specifically for a row, is only
incremented once per server, but decremented on every connection test
failure as part of homeUpdatePendingRequestsForConnectionTesting.