add global idle timeout (PIN lock)
It's possible for anyone to adjust/disable the screensaver settings on
the Roku. This will allow us to set an global timeout if disabled or
status unknown.

FW 5.6+ (supporting roAppManager.GetScreensaverTimeout)
 * use screensaver timeout value
 * use 5 minutes if screen saver disabled

FW < 5.6
 * use 30 minutes (max screensaver idle time)

First to lock the screen wins. If the screensaver locks, the global idle
timeout will play nicely, as will the reverse.