fix possible loop waiting for servers learned through plex.tv
It's possible if GDM is disabled (server or client), and a user is
connected to a Plex Account having zero servers, that a loop will occur
waiting for our connection tests to occur. This would normally be OK as
we should have zero pending requests, but that global was never set. The
"pendingRequests" global is fixed, and we have added a
loopguard (maxDuration) for a safety net.

fixes plexinc/roku-client#164