Add screensaver/lock screen workaround for legacy Rokus.
We key off the screensaver to lock the screen. Legacy devices will show
the lock screen immediately, instead of keeping the screen saver in
view. This seems to also prevent the screensaver from running again,
keeping the lock screen displayed indefinitely. This fix (workaround)
will send an invalid keypress to the roku after the lockscreen is
displayed to reenable the internal roku idle timeouts, allowing the
screensaver to start again.